requirements.txt
asgiref==3.11.1
attrs==25.4.0
certifi==2026.2.25
dj-database-url==0.5.0
Django==5.2.12
djangorestframework==3.16.1
drf-spectacular==0.29.0
inflection==0.5.1
jplephem==2.24
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
numpy==2.4.2
psycopg2==2.9.11
psycopg2-binary==2.9.11
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sgp4==2.25
skyfield==1.54
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
